1_The_Physical/Guess_Point_Level_1_Stateless/2_Arrange_Act_Assert/2_1_Password_Validator/src/index.ts
type PasswordValidationError =
  | "PasswordTooShort"
  | "PasswordTooLong"
  | "DigitMissing";

export type PasswordValidationResult = {
  valid: boolean;
  errors: PasswordValidationError[];
};

export function validatePassword(password: string): PasswordValidationResult {
  const result: PasswordValidationResult = { valid: true, errors: [] };

  if (password.length < 5) {
    result.errors.push("PasswordTooShort");
  } else if (password.length > 15) {
    result.errors.push("PasswordTooLong");
  }

  if (!containsDigit(password)) {
    result.errors.push("DigitMissing");
  }

  if (result.errors.length > 0) {
    result.valid = false;
  }

  return result;
}

function containsDigit(str: string): boolean {
  const digitRegex = /\d/;
  return digitRegex.test(str);
}
